fix(status holding): Handle messages with status holding.
When a message is read from PI with status holding the end time is null.
Added handling for the holding status and allow null end times. null end times are sorted last in the message display.

Started changes to ApplicationProperties.java to allow configuration for the maximum number of messages per WS call as well as the message initial start date on first run of the application. ApplicationProperties.java is updated - now need to refactor zpi_ejb to use the values instead of constants.